pyproject.toml
[tool.poetry]
name = "aws_cdk_15"
version = "0.1.0"
description = ""
authors = ["velo.yamigiku <velo.yamigiku@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aws-cdk-lib = "^2.25.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
